Accept ring doesn't selects, only accepts
If radius < r < accept-radius, don't change selected. We shouldn't
set it to an int, because if the user is coming from the outside (e.g.
when the menu is started close to the edge), then we shouldn't select
and accept, but give the user a chance to go into the selection area and
select something else. We don't want to set it to None, as that would
erase the user's previous selection when using the accept ring.